Task3 P2/P2/requirements.txt
tensorflow
keras
numpy
pillow
streamlit